+++ CHANGELOG.md
@@ -8,6 +8,9 @@ All notable changes to this project will be documented in this file.
 ### Changed
 - PNDA-3098: Sort kafka topics by name PNDA console.
 
+### Added:
+- PNDA-2445: Support for Hortonworks HDP
+
 ## [0.2.0] 2017-06-29
 ### Added:
 - PNDA-2691: Refactor for offline installation